// LDC.Atlas.WebApp/Client/app/shared/entities/counterparty-address.entity.ts
export interface CounterpartyAddress {
    randomId: number;
    counterpartyId: number;
    addressTypeID: number;
    addressTypeCode: string;
    addressId: number;
    address1: string;
    address2: string;
    city: string;
    zipCode: string;
    provinceID: number;
    countryID: number;
    mail: string;
    phoneNo: number;
    faxNo: number;
    addresTypeName: string;
    ldcRegionId: number;
    ldcRegionCode: string;
    isDeactivated: boolean;
    main: boolean;
    isDeleted: boolean;
}